Refactor: Flatten modules and simplify structure
- Flatten `src/errors/` directory to `src/errors.rs`.
- Merge `src/parser/grammar.rs` and `src/parser/recursion.rs` into `src/parser/mod.rs`.
- Merge `src/semantic/types.rs` into `src/semantic/model.rs`.
- Merge `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`.
- Update imports and documentation.
- Update `.jules/razor.md` with reductions.
